further clean up of proxy bridge
