+++ src/worker/workerbridge.ts
@@ -11,12 +11,10 @@ Paradigms:
 
 import { DBG_LOG, DBG_ME, DBG_OTHER } from "./debug"
 import { handleMessage } from "./handlers"
-import {ID, Key, makeLazyPath, Reference} from "./path"
-import { BridgeMsg, CallMsg, DisposeMsg, LinkMsg, NewMsg, RegisterMsg, SetMsg } from "./messages"
+import { Key } from "./path"
+import { BridgeMsg } from "./messages"
+import { _createProxy } from "./proxy"
 
-type Pod = any
-
-type TaggedCallback = Function & {__cb__?:number}
 
 export class WorkerBridge{
 	
@@ -34,6 +32,8 @@ export class WorkerBridge{
 	// remotely manually registered objects
 	remoteRegistry = new Set<string>()
 
+	// proxies that were requested but we don't know yet if the other side has registered them yet
+	// they will hang in promise limbo until we get the respective "reg" message
 	pendingProxies:{
 		[key: string]: (()=>void)[] // array of promise resolver functions
 	} = {}
@@ -68,6 +68,7 @@ export class WorkerBridge{
 	}
 
 	msgQueue: BridgeMsg[] = []
+
 	enqueueMsg(msg:BridgeMsg){
 		this.msgQueue.push(msg)
 	}
@@ -123,253 +124,21 @@ export class WorkerBridge{
 			throw new Error("numeric indices are reserved for anonymous objects")
 		}
 		if(remoteKey in this.remoteRegistry){
-			return this._createProxy(remoteKey)
+			return _createProxy(this, remoteKey)
 		}
 		else{
 			let bridge = this
-			return new Promise(function(resolve, reject){
+			return new Promise(function(resolve){
 				if(!(remoteKey in bridge.pendingProxies))
 				{
 					bridge.pendingProxies[remoteKey] = []
 				}
 				bridge.pendingProxies[remoteKey].push(
 					function(){
-						resolve(bridge._createProxy(remoteKey))
+						resolve(_createProxy(bridge, remoteKey))
 					}
 				)
 			})
-		}
-	}
-
-	// more general proxy creation, also for anonymous objects and child objects
-	_createProxy(ref: Reference):any{
-		// ref defines how this proxy relates to its target on the other side
-		// - ID as Key means, the proxied object was explicitly registered
-		//   on the remote side with a name
-		// - ID as Index means, the proxied object is the result of a get
-		//   or a call and it has no name but an index
-		// - LazyPath means that this proxy relates to its target by being a field
-		//   of its parent proxy
-
-		DBG_LOG(DBG_ME, "creates proxy to", 
-			JSON.stringify((typeof(ref) === "function")? ref() : ref))
-
-		// keep track of all the members that were get-ed
-		let children:{[key: string]:any}={}
-	
-		let bridge = this
-
-		// If not yet done, directly link this proxy
-		// to the corresponding object on the remote side.
-		// It gets an index entry in the remote registry.
-		function resolvePath():ID {
-			if(typeof(ref) === "function"){
-				DBG_LOG(
-					DBG_ME, "requests",
-					DBG_OTHER, "[", bridge.remoteOtherIndexCounter, "] =", JSON.stringify(ref())
-				)
-
-				bridge.enqueueMsg({kind:"link", path:ref()})
-				ref = bridge.remoteOtherIndexCounter--
-			}
-			return ref
-		}
-
-		// this will be a special member of the proxy that also
-		// deletes the registry entry on the remote side and
-		// sets the reference of this proxy to undefined
-		// so the GC can do its job
-		function dispose(){
-			bridge.enqueueMsg({
-				kind:"del",
-				id:resolvePath()
-			})
-			ref = undefined // ensure this proxy is not used further
-		}
-
-		return new Proxy(Object, {
-
-			set(target:any, prop:string, val:any){
-				bridge.enqueueMsg({
-					kind:"set",
-					id:resolvePath(),
-					prop, val:bridge.referencifyObject(val)
-				})
-				return true
-			},
-
-			get(target:any, prop:string, receiver:any){
-				if(prop === "then"){
-					return // nothing so this proxy cannot act as a promise
-				}
-				if(prop === "__ref__"){
-					return resolvePath()
-				}
-				if(prop === "dispose"){
-					return dispose
-				}
-				if(!(prop in children)){
-					children[prop] = bridge._createProxy(makeLazyPath(ref, prop))
-				}
-				return children[prop]
-			},
-
-			apply(target:any, thisArg:any, args:any[]){
-				bridge.enqueueMsg({
-					kind:"call",
-					id:resolvePath(),
-					args:bridge.referencifyObject(args)
-				})
-				DBG_LOG(
-					DBG_ME, "calls", DBG_OTHER, "[", resolvePath(), "]",
-					", store result in", DBG_OTHER, "[", bridge.remoteOtherIndexCounter, "]"
-				)
-				return bridge._createProxy(bridge.remoteOtherIndexCounter--)
-			},
-
-			construct(target:any, args:any){
-				bridge.enqueueMsg({
-					kind:"new",
-					id:resolvePath(),
-					args:bridge.referencifyObject(args)
-				})
-				DBG_LOG(
-					DBG_ME, "calls new on", DBG_OTHER, "[", resolvePath(), "]",
-					", store object in", DBG_OTHER, "[", bridge.remoteOtherIndexCounter, "]"
-				)
-				return bridge._createProxy(bridge.remoteOtherIndexCounter--)
-			}
-		})
-	}
-
-	referencifyFunction(fn:TaggedCallback){
-		if(fn.__cb__ === undefined){
-			// tag this function with a __cb__ index
-			// so that it will not be indexed multiple times
-			// TODO: if we allow the bridge to be reset, all tagged
-			// functions must be untagged
-			DBG_LOG(DBG_ME, "[", this.localOwnIndexCounter, "] = callback")
-			DBG_LOG(fn)
-
-			let newIndex = this.localOwnIndexCounter++
-			fn.__cb__ = newIndex
-			this.localRegistry[newIndex] = fn
-			return {__cb__:"new"}
-		}
-		else{
-			return {__cb__:fn.__cb__}
-		}
-	}
-
-	referencifyChildren(obj:any){
-		let copy:any = Array.isArray(obj)? [] : {}
-
-		for(let key in obj){
-
-			let wrapped = this.referencifyObject(obj[key])
-
-			if(typeof(wrapped) === "object"){
-				if("__cb__" in wrapped){
-					// this object or a sub object contains callbacks
-					copy["__cb__"] = true
-				}
-				if("__ref__" in wrapped){
-					// this object or a sub object contains proxies
-					copy["__ref__"] = true
-				}
-			}
-
-			copy[key] = wrapped
-		}
-		return copy
-	}
-
-	// scan an object for functions, store local references for them
-	// and replace the functions with the references
-	// this is for sending callbacks through the worker bridge
-	referencifyObject(value:any):Pod{
-		switch(typeof(value)){
-			case "number":
-			case "string":
-			case "boolean":
-				return value
-			case "object":
-				return this.referencifyChildren(value)
-			case "function":
-				if(value.__ref__ === undefined){
-					return this.referencifyFunction(value)
-				}
-				else{
-					// value is not a function but a proxy object
-					// (all bridge proxies wrap functions so operator() works)
-					return {__ref__:value.__ref__}
-				}
-		}
-
-		throw new Error("unsendable type: " + typeof(value))
-	}
-
-	resolveReferenceMembers(obj:any){
-		if("__cb__" in obj){
-			if(typeof(obj["__cb__"]) === "number"){
-				// reference to a callback that was already registered before
-				// return another proxy to it
-				// TODO: are we allowed to hand out multiple proxies to the same object?
-				// what if only one of them gets disposed?
-				return this._createProxy(obj["__cb__"])
-			}
-			else if(obj["__cb__"] === "new"){
-				DBG_LOG(
-					DBG_ME, "expects that",
-					DBG_OTHER, "[", this.remoteOwnIndexCounter, "] is a callback"
-				)
-				
-				// return a new proxy to a newly registered callback
-				return this._createProxy(this.remoteOwnIndexCounter++)
-			}
-			else if(obj["__cb__"] === true){
-				// a child or a grandchild etc. is a callback
-				for(let field in obj){
-					obj[field] = this.resolveReferences(obj[field])
-				}
-				return obj
-			}
-			else{
-				throw new Error("unexpected callback reference type")
-			}
-		}
-		else if("__ref__" in obj){
-			if(
-				typeof(obj["__ref__"]) === "number" ||
-				typeof(obj["__ref__"]) === "string"
-			){
-				// the remote object is a proxy, return the registry entry
-				return this.localRegistry[obj.__ref__]
-			}
-			else if(obj["__ref__"] === true){
-				// a child or a grandchild etc. is a proxy
-				for(let field in obj){
-					obj[field] = this.resolveReferences(obj[field])
-				}
-				return obj
-			}
-			else{
-				throw new Error("unexpected reference type")
-			}
-		}
-		else{
-			return obj
-		}
-	}
-
-	resolveReferences(value:Pod):any{
-		switch(typeof(value)){
-			case "number":
-			case "string":
-			case "boolean":
-				return value
-			case "object":
-				return this.resolveReferenceMembers(value)
 		}
 	}
 }
